perf: disable database cleanups when page is not active
